More documentation.

+++ PBJS/PBBezier.PathCommand.swift
@@ -10,16 +10,16 @@ import PBAPI
 extension PBBezier.PathCommand: JsParsable {
   
   public static let jsRules: [JsParseRule<Self>] = [
-    .a("move", [CGFloat.self, CGFloat.self], {
+    .a("move", [Float.self, Float.self], {
       try .move(to: CGPoint(x: $0.x(1) as CGFloat, y: $0.x(2)))
     }),
-    .a("line", [CGFloat.self, CGFloat.self], {
+    .a("line", [Float.self, Float.self], {
       try .addLine(to: CGPoint(x: $0.x(1) as CGFloat, y: $0.x(2)))
     }),
-    .a("scale", [CGFloat.self, CGFloat.self], {
+    .a("scale", [Float.self, Float.self], {
       try .apply(.identity.scaledBy(x: $0.x(1) as CGFloat, y: $0.x(2)))
     }),
-    .a("curve", [CGFloat.self, CGFloat.self], optional: [CGFloat.self], {
+    .a("curve", [Float.self, Float.self], optional: [Float.self], {
       try .addWeightedCurve(to: CGPoint(x: $0.x(1) as CGFloat, y: $0.x(2)), weight: $0.xq(3) ?? 0)
     })
   ]
